requirements.txt
torch
torchvision
opencv-python
numpy
Pillow
matplotlib
scikit-learn
streamlit
torch
torchvision
pytesseract
